+++ Tests/Packet++Test/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_executable(
   Packet++Test
   main.cpp
-  Tests/Asn1BerTests.cpp
+  Tests/Asn1Tests.cpp
   Tests/BgpTests.cpp
   Tests/CotpTests.cpp
   Tests/DhcpTests.cpp
