+++ Tests/Packet++Test/Tests/Asn1Tests.cpp
@@ -1,10 +1,10 @@
 #include "../TestDefinition.h"
-#include "Asn1BerCodec.h"
+#include "Asn1Codec.h"
 #include "RawPacket.h"
 #include "GeneralUtils.h"
 #include <functional>
 
-PTF_TEST_CASE(Asn1BerDecodingTest)
+PTF_TEST_CASE(Asn1DecodingTest)
 {
 	// Context specific
 	{
@@ -245,29 +245,29 @@ PTF_TEST_CASE(Asn1BerDecodingTest)
 	{
 		uint8_t data[20];
 		auto dataLen = pcpp::hexStringToByteArray("1f8100076d7976616c7565", data, 20);
-		PTF_ASSERT_RAISES(pcpp::Asn1Record::decode(data, dataLen), std::invalid_argument, "ASN.1 BER tags with value larger than 127 are not supported");
+		PTF_ASSERT_RAISES(pcpp::Asn1Record::decode(data, dataLen), std::invalid_argument, "ASN.1 tags with value larger than 127 are not supported");
 	}
 
 	// Not enough data to parse tag
 	{
 		uint8_t data[20];
 		auto dataLen = pcpp::hexStringToByteArray("1f8100076d7976616c7565", data, 20);
-		PTF_ASSERT_RAISES(pcpp::Asn1Record::decode(data, 0), std::invalid_argument, "Cannot decode ASN.1 BER record tag");
-		PTF_ASSERT_RAISES(pcpp::Asn1Record::decode(data, 1), std::invalid_argument, "Cannot decode ASN.1 BER record tag");
+		PTF_ASSERT_RAISES(pcpp::Asn1Record::decode(data, 0), std::invalid_argument, "Cannot decode ASN.1 record tag");
+		PTF_ASSERT_RAISES(pcpp::Asn1Record::decode(data, 1), std::invalid_argument, "Cannot decode ASN.1 record tag");
 	}
 
 	// Not enough data to parse length
 	{
 		uint8_t data[20];
 		pcpp::hexStringToByteArray("0500", data, 20);
-		PTF_ASSERT_RAISES(pcpp::Asn1Record::decode(data, 1), std::invalid_argument, "Cannot decode ASN.1 BER record length");
+		PTF_ASSERT_RAISES(pcpp::Asn1Record::decode(data, 1), std::invalid_argument, "Cannot decode ASN.1 record length");
 	}
 
 	// Incomplete record - doesn't contain the entire value
 	{
 		uint8_t data[20];
 		pcpp::hexStringToByteArray("0a022000", data, 20);
-		PTF_ASSERT_RAISES(pcpp::Asn1Record::decode(data, 3), std::invalid_argument, "Cannot decode ASN.1 BER record, data doesn't contain the entire record");
+		PTF_ASSERT_RAISES(pcpp::Asn1Record::decode(data, 3), std::invalid_argument, "Cannot decode ASN.1 record, data doesn't contain the entire record");
 	}
 
 	// Cast as the wrong type
@@ -277,10 +277,10 @@ PTF_TEST_CASE(Asn1BerDecodingTest)
 		auto record = pcpp::Asn1Record::decode(data, dataLen);
 		PTF_ASSERT_RAISES(record->castAs<pcpp::Asn1BooleanRecord>(), std::runtime_error, "Cast failed, instance isn't of the requested type");
 	}
-}; // Asn1BerDecodingTest
+}; // Asn1DecodingTest
 
 
-PTF_TEST_CASE(Asn1BerEncodingTest)
+PTF_TEST_CASE(Asn1EncodingTest)
 {
 	// Generic record
 	{
@@ -515,4 +515,4 @@ PTF_TEST_CASE(Asn1BerEncodingTest)
 		PTF_ASSERT_EQUAL(encodedValue.size(), dataLen);
 		PTF_ASSERT_BUF_COMPARE(encodedValue.data(), data, dataLen);
 	}
-} // Asn1BerEncodingTest
+} // Asn1EncodingTest
